Fix scala version issues for Invariant[Numeric].
Also add instances to Invariant companion and instances.all.
