+++ core/src/main/scala-2.13+/cats/instances/all.scala
@@ -25,6 +25,7 @@ trait AllInstances
     with FutureInstances
     with HashInstances
     with InvariantMonoidalInstances
+    with InvariantInstances
     with LazyListInstances
     with ListInstances
     with MapInstances
