+++ core/src/main/scala/cats/instances/invariant.scala
@@ -53,19 +53,8 @@ trait InvariantMonoidalInstances {
 }
 
 private[instances] trait InvariantInstances {
-
-  implicit val catsInvariantNumeric: Invariant[Numeric] = new Invariant[Numeric] {
-    def imap[A, B](fa: Numeric[A])(f: A => B)(g: B => A): Numeric[B] = new Numeric[B] {
-      def compare(x: B, y: B): Int = fa.compare(g(x), g(y))
-      def plus(x: B, y: B): B = f(fa.plus(g(x), g(y)))
-      def minus(x: B, y: B): B = f(fa.minus(g(x), g(y)))
-      def times(x: B, y: B): B = f(fa.times(g(x), g(y)))
-      def negate(x: B): B = f(fa.negate(g(x)))
-      def fromInt(x: Int): B = f(fa.fromInt(x))
-      def toInt(x: B): Int = fa.toInt(g(x))
-      def toLong(x: B): Long = fa.toLong(g(x))
-      def toFloat(x: B): Float = fa.toFloat(g(x))
-      def toDouble(x: B): Double = fa.toDouble(g(x))
-    }
+  implicit val catsInvariantForNumeric: Invariant[Numeric] = new Invariant[Numeric] {
+    def imap[A, B](fa: Numeric[A])(f: A => B)(g: B => A): Numeric[B] =
+      new ScalaVersionSpecificNumeric[A, B](fa)(f)(g) {}
   }
 }
